requirements.txt
markdown
jinja2
livereload
python-frontmatter
pygments
pymdown-extensions
markdown_link_attr_modifier
